+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-(No unreleased changes.)
+### Changed
+
+- Update `data/projects.ts` project case studies, tech stacks, and roles for MapBylaw, AlwaysGeeky Games, Emplifi, and ControlTech.
 
 ## [1.0.1] - 2026-02-26